@@ -2,9 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
 
 class BluetoothOffScreen extends StatelessWidget {
-  const BluetoothOffScreen({Key key, this.state}) : super(key: key);
-
-  final BluetoothState state;
+  const BluetoothOffScreen({Key? key, required this.state}) : super(key: key);
+  final BluetoothState? state;
 
   @override
   Widget build(BuildContext context) {
